llm_agent_client/requirements.txt
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.29.0
flask==3.0.0
uvicorn[standard]==0.24.0
